requirements.txt
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
Django==5.2.7
django-cors-headers==4.9.0
django-elasticsearch-dsl==7.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
elastic-transport==8.17.1
elasticsearch==7.17.12
elasticsearch-dsl==7.4.1
exceptiongroup==1.3.0
idna==3.11
mysqlclient==2.2.7
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==1.26.20
